apply changes suggested by reviewer
