+++ PolishWithBlondes/src/components/YoutubeList/YoutubeList.scss
@@ -16,9 +16,6 @@
   overflow: hidden;
   border-radius: 6px;
   box-shadow: 0 0.5rem 1rem rgba(0, 0, 0, 0.15);
-}
-
-.video-thumbnail-container {
   cursor: pointer;
 }
 
@@ -27,8 +24,9 @@
   width: 50px;
   height: 50px;
   z-index: 1;
-  left: 45%;
-  top: 40%;
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
 }
 
 .video-thumbnail {
@@ -42,7 +40,6 @@
 .video-title {
   font-family: var(--font-family);
   font-size: var(--text-font-size);
-  font-weight: 400;
   text-align: center;
   padding: 20px 0px 20px 0px;
 }
