--- PolishWithBlondes/src/components/home-info/home-info.scss
@@ -1,44 +1,0 @@
-@media (max-width: 1000px) {
-  .content-container.info {
-    height: 600px;
-  }
-}
-@media (max-width: 828px) {
-  .content-container.info {
-    height: 1200px;
-  }
-}
-
-//IN CASE I WANT TO GO BACK TO PREVIOUS METHOD OF HAVING SEPARATE CLASSES WITH THE SAME STYLES
-// .info-outer-container {
-//   background-color: var(--secondary-background);
-//   padding: 70px 0px;
-// }
-
-// .info-container {
-//   max-width: var(--max-width-container);
-//   flex-direction: column;
-//   margin: 0 auto;
-//   justify-content: center;
-// }
-
-// .info-title-container {
-//   max-width: var(--max-width-container);
-//   margin: 30px auto 60px 20px;
-//   font-family: var(--font-family);
-//   font-size: var(--header-font-size);
-//   font-weight: 600;
-// }
-// .info-title-container p {
-//   margin-bottom: 10px;
-// }
-// .line {
-//   width: 87px;
-//   border: 2px solid #3897d3;
-// }
-
-// .info-content-container {
-//   display: flex;
-//   flex-wrap: wrap;
-//   justify-content: space-around;
-// }
