--- PolishWithBlondes/src/components/home-youtube/home-youtube.scss
@@ -1,14 +1,0 @@
-.yt {
-  justify-content: space-between;
-}
-
-// @media (max-width: 1000px) {
-//   .content-container.yt {
-//     height: 600px;
-//   }
-// }
-// @media (max-width: 828px) {
-//   .content-container.yt {
-//     height: 1200px;
-//   }
-// }
