+++ PolishWithBlondes/src/components/info-section/InfoSection.scss
@@ -1,0 +1,10 @@
+@media (max-width: 1000px) {
+  .content-container.info {
+    height: 600px;
+  }
+}
+@media (max-width: 828px) {
+  .content-container.info {
+    height: 1200px;
+  }
+}
