+++ PolishWithBlondes/src/components/info-section/info-card/InfoCard.scss
@@ -37,7 +37,6 @@
   font-family: var(--font-family);
   font-size: var(--text-font-size);
   font-weight: 300;
-  // animation: fadeIn 1.5s;
   padding: 10px 0px 20px 0px;
 }
