+++ PolishWithBlondes/src/components/info-section/info-card/InfoCard.tsx
@@ -1,4 +1,4 @@
-import './info-card.scss';
+import './InfoCard.scss';
 
 export const InfoCard = ({
   title,
