+++ PolishWithBlondes/src/components/youtube-section/YoutubeSection.scss
@@ -1,0 +1,3 @@
+.yt {
+  justify-content: space-between;
+}
