+++ PolishWithBlondes/src/styles/global.scss
@@ -1,7 +1,3 @@
-* {
-  border: 1px solid red;
-}
-
 :root {
   --primary-font-color: #212529; // Dark gray for text contrast
